+++ R/descr.R
@@ -12,7 +12,7 @@
 #'    \code{max.length} chars.
 #' @param show Character vector, indicating which information (columns) that describe
 #'   the data should be returned. May be one or more of \code{"type", "label", "n",
-#'   "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "skew"}. There are
+#'   "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "iqr", "skew"}. There are
 #'   two shortcuts: \code{show = "all"} (default) shows all information,
 #'   \code{show = "short"} just shows n, missing percentage, mean and standard
 #'   deviation.
@@ -20,7 +20,7 @@
 #'    to console (\code{out = "txt"}) or as HTML-table in the viewer-pane
 #'    (\code{out = "viewer"}) or browser (\code{out = "browser"}).
 #'
-#' @inheritParams to_factor
+#' @inheritParams to_dummy
 #' @inheritParams frq
 #'
 #' @return A data frame with basic descriptive statistics.
@@ -78,7 +78,7 @@ descr <- function(x,
   # select elements that should be shown
 
   if ("all" %in% show)
-    show <- c("type", "label", "n", "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "skew")
+    show <- c("type", "label", "n", "NA.prc", "mean", "sd", "se", "md", "trimmed", "range", "iqr", "skew")
   else if ("short" %in% show)
     show <- c("n", "NA.prc", "mean", "sd")
 
@@ -90,10 +90,16 @@ descr <- function(x,
 
   w.name <- deparse(substitute(weights))
 
-  if (w.name == "NULL") {
-    w.name <- NULL
+  if (w.name != "NULL") {
+    w.name <- gsub("\"", "", w.name, fixed = FALSE)
+    if (!is.null(x[[w.name]])) {
+      dd[[w.name]] <- NULL
+      dd$.weights <- x[[w.name]]
+    } else {
+      dd$.weights <- eval(substitute(weights))
+    }
   } else {
-    dd$.weights <- x[[w.name]]
+    w.name <- NULL
   }
 
 
@@ -147,7 +153,7 @@ descr <- function(x,
 
 #' @importFrom dplyr select_if group_by summarise_all funs summarise
 #' @importFrom sjlabelled get_label
-#' @importFrom stats var na.omit sd median weighted.mean
+#' @importFrom stats var na.omit sd median weighted.mean IQR
 #' @importFrom rlang .data
 descr_helper <- function(dd, max.length) {
 
@@ -197,6 +203,7 @@ descr_helper <- function(dd, max.length) {
               as.character(round(min(.data$val, na.rm = TRUE), 2)),
               as.character(round(max(.data$val, na.rm = TRUE), 2))
             ),
+            iqr = stats::IQR(.data$val, na.rm = TRUE),
             skew = sjmisc.skew(.data$val)
           ))
     ) %>%
@@ -212,7 +219,7 @@ descr_helper <- function(dd, max.length) {
         dplyr::group_by(.data$var) %>%
         dplyr::summarise(
           n = round(sum(.data$.weights[!is.na(.data$val)], na.rm = TRUE)),
-          NA.prc = 100 * sum(is.na(.data$val)) / sum(.data$.weights[!is.na(.data$val)], na.rm = TRUE),
+          NA.prc = 100 * sum(is.na(.data$val)) / length(.data$val),
           mean = stats::weighted.mean(.data$val, w = .data$.weights, na.rm = TRUE),
           sd = wtd_sd_helper(.data$val, weights = .data$.weights),
           se = wtd_se_helper(.data$val, weights = .data$.weights),
@@ -221,7 +228,9 @@ descr_helper <- function(dd, max.length) {
             as.character(round(diff(range(.data$val, na.rm = TRUE)), 2)),
             as.character(round(min(.data$val, na.rm = TRUE), 2)),
             as.character(round(max(.data$val, na.rm = TRUE), 2))
-          )
+          ),
+          iqr = stats::IQR(.data$val, na.rm = TRUE),
+          skew = sjmisc.skew(.data$val)
         )
     ) %>%
       as.data.frame()
@@ -240,20 +249,6 @@ descr_helper <- function(dd, max.length) {
 }
 
 
-sjmisc.skew <- function(x) {
-  if (any(ina <- is.na(x)))
-    x <- x[!ina]
-
-  n <- length(x)
-  x <- x - mean(x)
-
-  if (n < 3)
-    return(NA)
-
-  sqrt(n) * sum(x ^ 3) / (sum(x ^ 2) ^ (3 / 2)) * sqrt(n * (n - 1)) / (n - 2)
-}
-
-
 wtd_se_helper <- function(x, weights) {
   sqrt(wtd_var(x, weights) / length(stats::na.omit(x)))
 }
@@ -268,9 +263,28 @@ wtd_var <- function(x, w) {
   x[is.na(w)] <- NA
   w[is.na(x)] <- NA
 
-  w <- na.omit(w)
-  x <- na.omit(x)
+  w <- stats::na.omit(w)
+  x <- stats::na.omit(x)
 
   xbar <- sum(w * x) / sum(w)
   sum(w * ((x - xbar)^2)) / (sum(w) - 1)
+}
+
+
+
+
+sjmisc.skew <- function(x) {
+  if (any(ina <- is.na(x)))
+    x <- x[!ina]
+
+  n <- length(x)
+  x <- x - mean(x)
+
+  if (n < 3)
+    return(NA)
+
+  # type-1 skewness
+  out <- (sum((x - mean(x))^3) / n) / (sum((x - mean(x))^2) / n)^1.5
+  # type-2 skewness
+  out * sqrt(n * (n - 1)) / (n - 2)
 }