+++ R/dicho.R
@@ -18,7 +18,7 @@
 #'    attributes of dichotomized variable (see \code{\link[sjlabelled]{set_labels}}).
 #'    If \code{NULL} (default), no value labels will be set.
 #'
-#' @inheritParams to_factor
+#' @inheritParams to_dummy
 #' @inheritParams rec
 #'
 #' @return \code{x}, dichotomized. If \code{x} is a data frame,
@@ -255,7 +255,7 @@ recode_fun <- function(x, .dat, fun, suffix, append, ...) {
 
       # names of grouping variables
       vars <- colnames(.dat)[colnames(.dat) %nin% grp.vars]
-      .dat <- dplyr::ungroup(.dat)
+      .dat <- as.data.frame(.dat)
 
       # iterate all groups
       for (i in unique(grps)) {