+++ R/re-exports.R
@@ -1,0 +1,18 @@
+#' @importFrom sjlabelled to_character
+#' @export
+sjlabelled::to_character
+
+
+#' @importFrom sjlabelled to_label
+#' @export
+sjlabelled::to_label
+
+
+#' @importFrom sjlabelled to_numeric
+#' @export
+sjlabelled::to_numeric
+
+
+#' @importFrom sjlabelled to_factor
+#' @export
+sjlabelled::to_factor