+++ R/remove_vars.R
@@ -8,9 +8,9 @@
 #' @param ... Character vector with variable names, or unquoted names
 #'   of variables that should be removed from the data frame.
 #'   You may also use functions like \code{:} or tidyselect's
-#'   \code{\link[tidyselect]{select_helpers}}.
+#'   select-helpers.
 #'
-#' @inheritParams to_factor
+#' @inheritParams to_dummy
 #'
 #' @return \code{x}, with variables specified in \code{...} removed.
 #'