+++ R/replace_na.R
@@ -15,7 +15,7 @@
 #'          that will be replaced by \code{value}. Herewith it is possible
 #'          to replace only specific \code{NA} values of \code{x}.
 #'
-#' @inheritParams to_factor
+#' @inheritParams to_dummy
 #' @inheritParams rec
 #'
 #' @return \code{x}, where \code{NA}'s are replaced with \code{value}. If \code{x}
@@ -59,31 +59,31 @@
 #' # show variables, NA's replaced with 99
 #' lapply(replace_na(dummy, v2, v3, value = 99), table, useNA = "always")
 #'
-#' library(haven)
-#' x <- labelled(c(1:3, tagged_na("a", "c", "z"), 4:1),
-#'               c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
-#'                 "Refused" = tagged_na("a"), "Not home" = tagged_na("z")))
-#' # get current NA values
-#' x
-#' get_na(x)
+#' if (require("haven")) {
+#'   x <- labelled(c(1:3, tagged_na("a", "c", "z"), 4:1),
+#'                 c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
+#'                   "Refused" = tagged_na("a"), "Not home" = tagged_na("z")))
+#'   # get current NA values
+#'   x
+#'   get_na(x)
 #'
-#' # replace only the NA, which is tagged as NA(c)
-#' replace_na(x, value = 2, tagged.na = "c")
-#' get_na(replace_na(x, value = 2, tagged.na = "c"))
+#'   # replace only the NA, which is tagged as NA(c)
+#'   replace_na(x, value = 2, tagged.na = "c")
+#'   get_na(replace_na(x, value = 2, tagged.na = "c"))
 #'
-#' table(x)
-#' table(replace_na(x, value = 2, tagged.na = "c"))
+#'   table(x)
+#'   table(replace_na(x, value = 2, tagged.na = "c"))
 #'
-#' # tagged NA also works for non-labelled class
-#' # init vector
-#' x <- c(1, 2, 3, 4)
-#' # set values 2 and 3 as tagged NA
-#' x <- set_na(x, na = c(2, 3), as.tag = TRUE)
-#' # see result
-#' x
-#' # now replace only NA tagged with 2 with value 5
-#' replace_na(x, value = 5, tagged.na = "2")
-#'
+#'   # tagged NA also works for non-labelled class
+#'   # init vector
+#'   x <- c(1, 2, 3, 4)
+#'   # set values 2 and 3 as tagged NA
+#'   x <- set_na(x, na = c(2, 3), as.tag = TRUE)
+#'   # see result
+#'   x
+#'   # now replace only NA tagged with 2 with value 5
+#'   replace_na(x, value = 5, tagged.na = "2")
+#' }
 #' @export
 replace_na <- function(x, ..., value, na.label = NULL, tagged.na = NULL) {
   # check for valid value