+++ tests/testthat/test-splitvar.R
@@ -1,17 +1,15 @@
-context("sjmisc, split_var")
+if (require("testthat") && require("sjmisc") && require("dplyr")) {
+  data(mtcars)
 
-library(sjmisc)
-library(dplyr)
-data(mtcars)
-
-test_that("std, split_var", {
-  tmp <- split_var(mtcars, disp, n = 3, append = FALSE)
-  expect_equal(sum(tmp$disp_g == 1), 11)
+  test_that("std, split_var", {
+    tmp <- split_var(mtcars, disp, n = 3, append = FALSE)
+    expect_equal(sum(tmp$disp_g == 1), 11)
 
 
-  tmp <- mtcars %>%
-    dplyr::group_by(cyl) %>%
-    split_var(disp, n = 3, append = FALSE)
+    tmp <- mtcars %>%
+      dplyr::group_by(cyl) %>%
+      split_var(disp, n = 3, append = FALSE)
 
-  expect_equal(sum(tmp$disp_g == 1), 10)
-})
+    expect_equal(sum(tmp$disp_g == 1), 10)
+  })
+}